AudioSessionManager 내 AudioPlayerAgent 의존성 제거
